+++ src/pages/Browse/Browse.css
@@ -165,70 +165,3 @@
 .dropdown-menu a:hover {
   background-color: #f4f4f4;
 }
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-bottom: 20px;
-}
-
-th,
-td {
-  padding: 10px;
-  text-align: left;
-  border-bottom: 1px solid #ccc;
-}
-
-th {
-  background-color: #f2f2f2;
-}
-
-
-input[type='text'],
-input[type='password'],
-input[type='email'] {
-  margin-bottom: 10px;
-  padding: 5px;
-  width: 200px;
-}
-
-button {
-  padding: 5px 10px;
-  background-color: #007bff;
-  color: #fff;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-button:hover {
-  background-color: #0056b3;
-}
-.actions-button {
-  display: flex;
-  justify-content:space-around;
-  align-items: center;
-
-}
-
-.actions-button button {
-  margin-right: 5px;
-}
-
-.add-row-button {
-  margin-top: 10px;
-  padding: 5px 10px;
-  background-color: #007bff;
-  color: #fff;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.add-row-button:hover {
-  background-color: #0056b3;
-}
-
-.edit-button {
-  background-color: #ffc107;
-  color: #000;
-}
