+++ src/pages/Browse/Browse.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import Users from './Users';
 import './Browse.css';
 
 interface BrowseItem {
@@ -8,26 +9,11 @@ interface BrowseItem {
     content: string;
 }
 
-interface TableRow {
-    id: number;
-    name: string;
-    password: string;
-    email: string;
-}
-
 const Browse: React.FC = () => {
     const [selectedItem, setSelectedItem] = useState<BrowseItem | null>(null);
     const [sidebarWidth, setSidebarWidth] = useState<number>(200);
     const [isDragging, setIsDragging] = useState<boolean>(true);
     const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-    const [tableRows, setTableRows] = useState<TableRow[]>([]);
-    const [newRow, setNewRow] = useState<TableRow>({
-        id: 0,
-        name: '',
-        password: '',
-        email: '',
-    });
-    const [isEditMode, setIsEditMode] = useState(false);
 
     const closeDropdown = () => {
         setIsDropdownOpen(false);
@@ -54,43 +40,6 @@ const Browse: React.FC = () => {
             const newSidebarWidth = event.clientX;
             setSidebarWidth(newSidebarWidth);
         }
-    };
-
-    const handleDeleteRow = (id: number) => {
-        const updatedRows = tableRows.filter((row) => row.id !== id);
-        setTableRows(updatedRows);
-    };
-
-    const handleEditRow = (id: number) => {
-        const rowToEdit = tableRows.find((row) => row.id === id);
-        if (rowToEdit) {
-            setNewRow(rowToEdit);
-            handleDeleteRow(id);
-            setIsEditMode(true);
-        }
-    };
-
-    const handleSave = () => {
-        if (isEditMode) {
-            const updatedRows = [...tableRows, newRow];
-            setTableRows(updatedRows);
-        }
-        setNewRow({ id: 0, name: '', password: '', email: '' });
-        setIsEditMode(false);
-    };
-
-    const handleAddRow = () => {
-        if (isEditMode) {
-            handleSave();
-        } else {
-            const newRowWithId = {
-                ...newRow,
-                id: Math.floor(Math.random() * 1000000),
-            };
-            setTableRows([...tableRows, newRowWithId]);
-        }
-        setNewRow({ id: 0, name: '', password: '', email: '' });
-        setIsEditMode(false);
     };
 
     const browseItems: BrowseItem[] = [
@@ -177,122 +126,7 @@ const Browse: React.FC = () => {
                             </div>
                         </div>
                         <p className="content-class">{selectedItem.content}</p>
-                        {selectedItem.label === 'Users' && (
-                            <div>
-                                <div>
-                                    <table>
-                                        <thead>
-                                            <tr>
-                                                <th>ID</th>
-                                                <th>Name</th>
-                                                <th>Password</th>
-                                                <th>Email</th>
-                                                <th>Actions</th>
-                                            </tr>
-                                        </thead>
-                                        <tbody>
-                                            {tableRows.map((row) => (
-                                                <tr key={row.id}>
-                                                    <td>{row.id}</td>
-                                                    <td>{row.name}</td>
-                                                    <td>{row.password}</td>
-                                                    <td>{row.email}</td>
-                                                    <td>
-                                                        <button
-                                                            onClick={() =>
-                                                                handleEditRow(
-                                                                    row.id
-                                                                )
-                                                            }
-                                                        >
-                                                            Edit
-                                                        </button>{' '}
-                                                        <button
-                                                            onClick={() =>
-                                                                handleDeleteRow(
-                                                                    row.id
-                                                                )
-                                                            }
-                                                        >
-                                                            Delete
-                                                        </button>
-                                                    </td>
-                                                </tr>
-                                            ))}
-                                            {isEditMode && (
-                                                <tr>
-                                                    <td>{newRow.id}</td>
-                                                    <td>
-                                                        <input
-                                                            type="text"
-                                                            placeholder="Name"
-                                                            value={newRow.name}
-                                                            onChange={(e) =>
-                                                                setNewRow({
-                                                                    ...newRow,
-                                                                    name: e
-                                                                        .target
-                                                                        .value,
-                                                                })
-                                                            }
-                                                        />
-                                                    </td>
-                                                    <td>
-                                                        <input
-                                                            type="password"
-                                                            placeholder="Password"
-                                                            value={
-                                                                newRow.password
-                                                            }
-                                                            onChange={(e) =>
-                                                                setNewRow({
-                                                                    ...newRow,
-                                                                    password:
-                                                                        e.target
-                                                                            .value,
-                                                                })
-                                                            }
-                                                        />
-                                                    </td>
-                                                    <td>
-                                                        <input
-                                                            type="email"
-                                                            placeholder="Email"
-                                                            value={newRow.email}
-                                                            onChange={(e) =>
-                                                                setNewRow({
-                                                                    ...newRow,
-                                                                    email: e
-                                                                        .target
-                                                                        .value,
-                                                                })
-                                                            }
-                                                        />
-                                                    </td>
-                                                    <td>
-                                                        <button
-                                                            onClick={handleSave}
-                                                        >
-                                                            Save
-                                                        </button>{' '}
-                                                        <button
-                                                            onClick={() =>
-                                                                setIsEditMode(
-                                                                    false
-                                                                )
-                                                            }
-                                                        >
-                                                            Cancel
-                                                        </button>
-                                                    </td>
-                                                </tr>
-                                            )}
-                                        </tbody>
-                                    </table>
-                                </div>
-                                <button onClick={handleAddRow}>Add User</button>
-                            </div>
-                        )}
+                        {selectedItem.label === 'Users' && <Users />}
                     </div>
                 ) : (
                     <div>No item Selected</div>
